Move all app-specific logic into the App class - even TileHighlight
TileHighlight should probably be part of the level editor and will not
be part of the main game, it doesn't belong in the Map.
Same for startpoint/endpoint/blocks and all the logic related to them -
they're specific to this little experiment and otherwise unrelated to
the Map.
The Map will finally be called MapView when it starts involving a Map
model.